pyproject.toml
[project]
name = "job-haunt-backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.120.1",
    "isort>=7.0.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "ruff[format]>=0.14.2",
]
